requirements.txt
click==6.7
Flask==0.12
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
olefile==0.44
Pillow==4.0.0
python-resize-image==1.1.11
requests==2.13.0
Werkzeug==0.11.15
